Decompose monolithic frontend modules and extract prediction services
Split app/src/charts.js (~1300 lines) into focused submodules under
app/src/charts/ (colors, core, overlays, solution-charts, ev-charts,
ev-annotations) and app/src/predictions.js into app/src/predictions/
(forecast-chart, forecast-series, config-form, accuracy-charts).
Extract EV panel logic into ev-settings.js and run orchestration into
optimizer-controller.js. On the server side, pull prediction adjustment
CRUD and forecast execution out of the predictions route into dedicated
service files. Add tests for all new browser modules.
