+++ AGENTS.md
@@ -5,17 +5,22 @@
 - The **server** owns the persisted state under `DATA_DIR` (defaults to `<repo>/data`).
   - `settings.json` holds system and algorithm scalars. Defaults live in `api/defaults/default-settings.json`.
   - `data.json` holds time series + SoC and starts from `api/defaults/default-data.json`.
-- Time-series data comes from the VRM API (via `vrm-refresh.js`) and is persisted server-side. The client only visualizes it.
-- The LP config (`lib/build-lp.js`) is derived from persisted settings + data; the client never sends LP parameters.
+  - `prediction-config.json` holds load/PV prediction model settings.
+- Time-series data comes from the VRM API (via `api/services/vrm-refresh.ts`) or `/data`/prediction endpoints when a data source is set to `api`. The client only visualizes it.
+- Manual prediction adjustments are stored in `data.json` and applied on forecast responses without changing the raw persisted forecast series.
+- The LP config (`lib/build-lp.ts`) is derived from persisted settings + data; the client never sends LP parameters.
 
 ## Front-end layout
 - Static UI lives in `app/index.html` and `app/main.js`.
 - Browser-side modules under `app/src/`:
 
   - `app/src/api/client.js` — low-level `getJson` / `postJson`.
-  - `app/src/api/api.js` — endpoint wrappers (settings, calculate, VRM refresh).
+  - `app/src/api/api.js` — endpoint wrappers (settings, calculate, VRM refresh, predictions, EV/HA helpers).
   - `app/src/config-store.js` — loads and saves the current settings snapshot via the API.
-  - `app/src/charts.js`, `app/src/table.js` — visualization only.
+  - `app/src/charts.js` — compatibility barrel for chart modules under `app/src/charts/`.
+  - `app/src/predictions.js` — Predictions tab coordinator; detailed form/chart logic lives under `app/src/predictions/`.
+  - `app/src/ev-settings.js`, `app/src/ev-tab.js` — EV settings wiring and EV tab visualization.
+  - `app/src/table.js` — schedule table visualization.
   - `app/src/utils.js` — small utilities (e.g. debounce).
 
 ### Settings on the client
@@ -26,6 +31,7 @@
 - `snapshotUI()` collects:
   - **system settings** (battery capacity, step size, grid/battery limits, …),
   - **algorithm settings** (terminal SoC mode, custom price, …),
+  - **data-source and EV settings**,
   - UI-only bits (e.g. `tableShowKwh`).
 
 Snapshots are saved via `POST /settings` when inputs change (debounced) and before a recompute.
@@ -37,8 +43,14 @@ All routes are implemented in `api/`. Important ones:
 - `GET /settings` — returns persisted settings or defaults when missing.
 - `POST /settings` — merges the incoming object onto existing settings and writes to `DATA_DIR/settings.json`.
 - `POST /calculate` — builds the LP from persisted settings + data, runs HiGHS, and returns rows/summary/diagnostics. Optional body flags: `updateData` (refresh VRM series before solving) and `writeToVictron` (attempt MQTT schedule write).
+- `GET /data`, `POST /data` — read or merge persisted time-series data.
 - `POST /vrm/refresh-settings` — refresh relatively static system limits/settings from VRM and persist.
+- `GET/POST /predictions/config`, `/predictions/validate`, `/predictions/*/forecast`, `/predictions/adjustments` — prediction config, validation, forecast generation, and manual adjustment CRUD.
+- `GET /ev/current`, `GET /ev/schedule` — current and full EV charging schedule from the last computed plan.
+- `GET /ha/entity/:entityId` — live Home Assistant entity lookup for settings validation.
+
+Prediction routing lives in `api/routes/predictions.ts`; orchestration and persistence helpers live in `api/services/prediction-forecast-runner.ts` and `api/services/prediction-adjustment-store.ts`.
 
 ## PR / testing notes
 - Prefer small, focused commits with descriptive messages.
-- Run `npm run lint` or relevant integration checks when modifying solver or API behaviour. Document executed commands in the final summary.
+- Run `npm run typecheck`, `npm run lint`, or relevant integration checks when modifying solver or API behaviour. Use `npm run test:run` for a one-shot test suite. Document executed commands in the final summary.
