+++ app/src/utils.js
@@ -1,3 +1,8 @@
+export function toDatetimeLocal(d) {
+  const pad = (n) => String(n).padStart(2, '0');
+  return `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())}T${pad(d.getHours())}:${pad(d.getMinutes())}`;
+}
+
 export function escapeHtml(value) {
   return String(value)
     .replaceAll('&', '&amp;')
